[Lydian] Support for KVM traffic in Lydian
For supporting KVM-ubuntu traffic in Lydian we need to ensure correct
installation of lydian service on KVM-ubuntu.

Once that is tackled, for running the traffic we need to make sure
endpoint maps of IPs and interfaces are correctly identified and
prepared on both runner and remote machines.

Following pdb shell log does the installation of lydian and runs
traffic in the end with 100% pass percent.
http://blr-dbc2108.eng.vmware.com/hipandey/lydian_kvm_support/kvm_support_pdb_shell.txt

Issue: #
Jira: #NPTF-3122
Change-Id: I0e88440069be607aa9b49c830a5450603f4257bb